requirements.txt
psycopg2
python-dotenv
flask
flask-cors